Add support for not replicating endpoint tables
For sites that should not be "multi region" in the sense that we have
had them before, we need to have a way to skip replication of endpoint
tables. The --no-endpoints flag controls this and is off by default.

Also updates the restore script to support running against a remote,
similar to the snapshot.