[keystone] Add federation configuration
This adds support for OpenID login by adding the "external" and "openid"
auth provider support. "External" is needed because under the hood the
Keystone Apache installation is passing auth information as environment
variables to the WSGI process; 'external' is the mechanism that can
marshall this information to Keystone and is needed for any type of
federation for this reason.