[prometheus] Make playbook less expensive
Turns out there is no need to run this huge loop that pulls down a
many-to-many mapping of addresses; we can just use hostvars, as all the
hosts are part of the same play.

Also fixes up issue in the snmp playbook where it would always register
a change on the `make mibs` task.