[letsencrypt] Add LetsEncrypt role
This role allows managing the certificates for HAProxy via LetsEncrypt.
It is assumed that this role be run on the same host as HAProxy.

*Note*: currently this assumes that only one node is running HAProxy! It
is not HA-compatible. For multiple nodes, we still need some sort of
mechanism to copy the certs to all hosts.

The way this will work:

  - 1. Bootstrap the servers with KA.
  - 2. Run the LetsEncrypt playbook to generate the certbot script.
  - 3. Run the certbot script to request the initial cert(s).
  - 4. Proceed with deploying HAProxy / the rest of KA.

If adding new certs to an existing deploy:

  - 1. Ensure letsencrypt_domains is empty.
  - 2. Run the LetsEncrypt playbook to generate the certbot script.
  - 3. Run the certbot script to request the initial cert(s).
  - 4. Update letsencrypt_domains to include the cert domains.
  - 5. Reconfigure/upgrade HAProxy

A periodic task will be added to renew the certificates automatically.